requirements.txt
aiogram==3.20.0.post0
python-dotenv==1.1.0
aio-pika==9.5.5
redis==6.2.0
pandas==2.2.3
xlsxwriter==3.2.3
motor==3.7.0